requirements.txt
matplotlib==3.7.5
numpy==1.24.3
opencv_python==4.10.0.84
pandas==2.0.3
Pillow==11.2.1
psutil==5.9.0
psutil==6.0.0
pyautogui==0.9.54
scikit_learn==1.3.2
scipy==1.15.3
seaborn==0.13.2
skimage==0.0
sympy==1.13.3
timm==1.0.15
torch==1.13.0
torchsummary==1.5.1
torchvision==0.14.0
tqdm==4.66.5
